New CFD method
This method tries to replicate an analog cfd in software. this is different from the TraditionalCfd which does not use a threshold to find the zero crossing (it uses the vector:: min(max)_element to find the crossing points) and uses another method to find the phase (this appears to be a traditional digital cfd) . 

This  method seems to work but is still being benchmarked. It appears to be at least functional for an online mode. 

This commit also changes the DefaultConfiguration values to be reasonable for this cfd. Right now this is ok, since an upgrade to the DefaultConfiguration values system is planned (and imminent) that will change the DefaultConfiguration from a name space to a class, which will allow us to have different default values based on type (or type:subtype)