New Default XML Parameter Structure
We now have up to type:subtype:group specific defaults, with type:subtype sets available, and the FINAL fallback being the old DefaultConfig namespace (DefaultConfigurationValues.hpp)

With this system each type will get its own header file. Where it is a derived class of the DefaultDetectorConfig class from NewDetectorConfigurationValues. 

I have included the big 3 when it comes to our analyses (pspmt, beta, vandle) as well as a Template. 

The Beta parameters are scavanged from old XMLs. Note the special set "beta:double:timing" this is for the ANL1471 experiment where we had the Sensl Fast SiPM output. These have hard coded fitting parameters (because the shape is a gaus). So I am not including dummy values in the Map.  The "double:ornl2016" group is for the Skutek 1x4 SiPM boards used at that experiment. The Waveforms for those were malformed. And finally the "double:isolde" fitting parameters are taken from the the is632 run at ISOLDE, since the beta there is provided by IDS.

The Pspmt values are pulled from the RIBF136 (RIKEN2018) experiment, they were also the same parameters used in the e14060 (NSCL2016) experiment.  I'm including fields for the implantation beam detectors (f11,ion (front plastic), veto, desi) for Fitting, Trace and Cfd because we took traces on these detectors, however they are currently ignored by the Waveform and Fitting Analyzers.

This framework is expandable in the same way that the EventProcessor is.
The addition of a new Set of defaults (says for the gscint type) would require the <Type>DefaultConfig.hpp (ex: GscintDefaultConfig.hpp) header, as well as an addition to the DefaultDetectorConfig class constructor.

The updates to the MapNodeXmlParser::ParseCfdNode() will come in the next commit, due to a change in how the Cfd parameters are stored. Which requires changes to the CfdAnalyzer Classes to take the new container. (Got ahead of my self when making these upgrades, and I don't want it to be such a large commit.)

Tests include "cout'ing" the parameters that are set for a given channel after parsing , and first order comparisons of scanned data for ORNL2016. All seems well. 

This system will also be  expanded as NEXT is brought into mainline PAASS.