Update to how we store CFD parameters
We now parse, store, and pass the software Cfd parameters as a Map. This makes the retrieval of the parameters less obscure, and also allows us to have an arbitrary number of available parameters, if needed, in the future. This also reduces the number of overloads we need for "CalculatePhase()" in the TimingDriver.hpp

Additionally we now have an extra parameter because there is not a safe / sane default for the 2 Cfd algorithms that are in use (PolyCfd) or being developed (PixieCfd).  The new parameter goes to the PixieCfd since that is the one currently being developed. cfdT is the threshold for that calculation, and is >1. However, cfdF is also a threshold but for the polyCfd. There the threshold is understood as a percentage of the extrapolated maximum of the pulse, and therefore must be 0<cfdF<1.