Fixing RD without root output
Previously RD had no protections for sysroot being off.
This would cause a segfault when it tried to dump the
"root_dev_vec_" up onto the "pixie_tree_event_".

The addition of the DetectorDriver header allows us to check
the "sysroot_" bool.

also i have moved the reset of the structure to the beginning
of the loop over RD events. I did this so that we can
guarantee that the structure is reset before each loop, while
also being fully init'd on the first go around.